energy-management-system/ml-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
influxdb-client>=1.44.0
scikit-learn>=1.4.2
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
requests>=2.31.0
